+++ Sources/MacEverything/MacEverythingApp.swift
@@ -1,15 +1,20 @@
-import AppKit
 import SwiftUI
+
+#if SWIFT_PACKAGE
+import AppKit
 
 final class AppDelegate: NSObject, NSApplicationDelegate {
     func applicationDidFinishLaunching(_ notification: Notification) {
         WindowActivation.activateApp()
     }
 }
+#endif
 
 @main
 struct MacEverythingApp: App {
+    #if SWIFT_PACKAGE
     @NSApplicationDelegateAdaptor(AppDelegate.self) private var appDelegate
+    #endif
 
     var body: some Scene {
         WindowGroup {